refactor: split pseudo-etf-charts.tsx into chart/pseudo-etf/ folder
Closes #361
