+++ frontend/src/components/chart/pseudo-etf/index.ts
@@ -1,0 +1,2 @@
+export { PerformanceOverlayChart } from "./performance-overlay-chart"
+export { DailyContributionChart } from "./daily-contribution-chart"
